+++ rainnow/src/plotting.py
@@ -9,6 +9,7 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 from cartopy.feature import NaturalEarthFeature
+import torch 
 
 
 def plot_raw_imerg_xarray_tile(
@@ -655,3 +656,91 @@ def plot_dyffusion_predictions(
 
     plt.tight_layout()
     plt.subplots_adjust(**layout_params)
+
+
+
+def plot_training_val_loss(
+    train_losses: List[float],
+    val_losses: List[float],
+    criterion_name: str,
+    figsize: Dict[str, Any] = (8, 6),
+) -> plt.Figure:
+    """
+    Plot the training and validation loss curves.
+
+    Parameters
+    ----------
+    train_losses : list or array-like
+        The training loss values for each epoch.
+    val_losses : list or array-like
+        The validation loss values for each epoch.
+    criterion_name : str
+        The name of the loss criterion (e.g., 'MSE', 'Cross Entropy').
+    figsize : tuple, optional
+        The figure size in inches (width, height). Default is (8, 6).
+
+    Returns
+    -------
+    fig : plt.Figure
+        The generated figure object containing the plot.
+    """
+
+    num_epochs = [i for i in range(len(train_losses))]
+    fig, ax = plt.subplots(figsize=figsize)
+    ax.plot(num_epochs, train_losses, color="C0", label="train loss")
+    ax.plot(num_epochs, val_losses, color="C1", label="val loss")
+    ax.set_ylabel(f"{criterion_name}")
+    ax.set_xlabel("epochs")
+    ax.legend(loc="best")
+
+    return fig
+
+
+def plot_predicted_sequence(
+    X: torch.Tensor,
+    target: torch.Tensor,
+    pred: torch.Tensor,
+    batch_num: int,
+    plot_params: Dict[str, Any],
+    figsize: Tuple[float, float],
+) -> None:
+    """
+    Plot the input sequence, target sequence, and predicted sequence.
+
+    This function creates a visualization of the input sequence, the target sequence,
+    and the predicted sequence for a specific batch.
+
+    Parameters
+    ----------
+    X : torch.Tensor
+        The input sequence tensor of shape (batch_size, input_sequence_len, channels, height, width).
+    target : torch.Tensor
+        The target sequence tensor of shape (batch_size, target_sequence_len, channels, height, width).
+    pred : torch.Tensor
+        The predicted sequence tensor of shape (batch_size, target_sequence_len, channels, height, width).
+    batch_num : int
+        The index of the batch to plot.
+    plot_params : Dict[str, Any]
+        A dictionary of parameters to pass to the imshow function.
+    figsize : Tuple[float, float]
+        The figure size in inches (width, height).
+
+    Returns
+    -------
+    None
+    """
+    input_sequence_len = X.size(1)
+    target_sequence_len = target.size(1)
+    nrows = 2
+    fig, axs = plt.subplots(nrows=nrows, ncols=input_sequence_len + target_sequence_len, figsize=figsize)
+    for i in range(input_sequence_len + target_sequence_len):
+        if i < input_sequence_len:
+            axs[0, i].imshow(X[batch_num, i, 0, :, :], **plot_params)
+        else:
+            axs[0, i].imshow(target[batch_num, i - input_sequence_len, 0, :, :], **plot_params)
+            axs[1, i].imshow(pred[batch_num, i - input_sequence_len, 0, :, :], **plot_params)
+
+    for ax in axs.flatten():
+        ax.axis("off")
+
+    plt.tight_layout()
